Создал пакет service, storage для genre, mpa и свои классы, доработал service storage, controller film, файлы sql и UserRowMapper

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -15,7 +15,7 @@ public interface FilmStorage {
 
     void deleteFilm(long id);
 
-    void addLike(int filmId, int userId);
+    void addLike(long filmId, long userId);
 
     void removeLike(long filmId, long userId);
